MR-93: form components
